# openapi.yaml
openapi: 3.0.3
info:
  title: QAIra API
  version: 1.0.0
  description: |
    OpenAPI specification for the current QAIra Fastify + PostgreSQL backend.
    This spec documents the implemented CRUD resources, request payloads,
    query filters, and the execution lifecycle actions.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /projects/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [Projects]
      summary: Get project by ID
      responses:
        "200":
          description: Project record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      tags: [Projects]
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedResponse"
    delete:
      tags: [Projects]
      summary: Delete project
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
...
  /app-types:
    get:
      tags: [App Types]
      summary: List app types
      parameters:
        - $ref: "#/components/parameters/ProjectIdQuery"
      responses:
        "200":
          description: List of app types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppType"
    post:
      tags: [App Types]
      summary: Create app type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppTypeRequest"
      responses:
        "200":
          description: App type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /app-types/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [App Types]
      summary: Get app type by ID
      responses:
        "200":
          description: App type record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppType"
    put:
      tags: [App Types]
      summary: Update app type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppTypeRequest"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedResponse"
    delete:
      tags: [App Types]
      summary: Delete app type
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
  /requirements:
    get:
      tags: [Requirements]
      summary: List requirements
      parameters:
        - $ref: "#/components/parameters/ProjectIdQuery"
        - $ref: "#/components/parameters/StatusQuery"
        - $ref: "#/components/parameters/PriorityQuery"
      responses:
        "200":
          description: List of requirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Requirement"
    post:
      tags: [Requirements]
      summary: Create requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequirementRequest"
      responses:
        "200":
          description: Requirement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /requirements/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [Requirements]
      summary: Get requirement by ID
      responses:
        "200":
          description: Requirement record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
    put:
      tags: [Requirements]
      summary: Update requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequirementRequest"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedResponse"
    delete:
      tags: [Requirements]
      summary: Delete requirement
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
...
  /test-cases:
    get:
      tags: [Test Cases]
      summary: List test cases
      parameters:
        - $ref: "#/components/parameters/SuiteIdQuery"
        - $ref: "#/components/parameters/RequirementIdQuery"
        - $ref: "#/components/parameters/StatusQuery"
      responses:
        "200":
          description: List of test cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCase"
    post:
      tags: [Test Cases]
      summary: Create test case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestCaseRequest"
      responses:
        "200":
          description: Test case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /test-cases/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [Test Cases]
      summary: Get test case by ID
      responses:
        "200":
          description: Test case record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
    put:
      tags: [Test Cases]
      summary: Update test case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestCaseRequest"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedResponse"
    delete:
      tags: [Test Cases]
      summary: Delete test case
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
...
  /executions:
    get:
      tags: [Executions]
      summary: List executions
      parameters:
        - $ref: "#/components/parameters/ProjectIdQuery"
        - $ref: "#/components/parameters/ExecutionStatusQuery"
      responses:
        "200":
          description: List of executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Execution"
    post:
      tags: [Executions]
      summary: Create execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExecutionRequest"
      responses:
        "200":
          description: Execution created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /executions/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [Executions]
      summary: Get execution by ID
      responses:
        "200":
          description: Execution record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
    delete:
      tags: [Executions]
      summary: Delete execution
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
  /executions/{id}/start:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    post:
      tags: [Executions]
      summary: Start execution
      responses:
        "200":
          description: Start result
          content:
            application/json:
              schema:
                type: object
                properties:
                  started:
                    type: boolean
                    example: true
  /executions/{id}/complete:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    post:
      tags: [Executions]
      summary: Complete execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteExecutionRequest"
      responses:
        "200":
          description: Completion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  completed:
                    type: boolean
                    example: true
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    ProjectIdQuery:
      name: project_id
      in: query
      schema:
        type: string
...
    SuiteIdQuery:
      name: suite_id
      in: query
      schema:
        type: string
    RequirementIdQuery:
      name: requirement_id
      in: query
      schema:
        type: string
...
    StatusQuery:
      name: status
      in: query
      schema:
        type: string
    PriorityQuery:
      name: priority
      in: query
      schema:
        type: integer
    ExecutionStatusQuery:
      name: status
      in: query
      schema:
        type: string
        enum: [queued, running, completed, failed]
...
  schemas:
    IdResponse:
      type: object
      properties:
        id:
          type: string
      required: [id]
    UpdatedResponse:
      type: object
      properties:
        updated:
          type: boolean
          example: true
      required: [updated]
    DeletedResponse:
      type: object
      properties:
        deleted:
          type: boolean
          example: true
      required: [deleted]
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        created_by: { type: string, nullable: true }
        created_at: { type: string }
    CreateProjectRequest:
      type: object
      required: [name, created_by]
      properties:
        name: { type: string }
        description: { type: string }
        created_by: { type: string }
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
...
    AppType:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [web, api, android, ios, unified]
        is_unified: { type: integer }
        created_at: { type: string }
    CreateAppTypeRequest:
      type: object
      required: [project_id, name, type]
      properties:
        project_id: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [web, api, android, ios, unified]
        is_unified: { type: boolean }
    UpdateAppTypeRequest:
      type: object
      properties:
        name: { type: string }
        is_unified: { type: boolean }
    Requirement:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        priority: { type: integer, nullable: true }
        status: { type: string, nullable: true }
        created_at: { type: string }
    CreateRequirementRequest:
      type: object
      required: [project_id, title]
      properties:
        project_id: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: integer }
        status: { type: string }
    UpdateRequirementRequest:
      type: object
      properties:
        project_id: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: integer }
        status: { type: string }
...
    TestCase:
      type: object
      properties:
        id: { type: string }
        suite_id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        priority: { type: integer, nullable: true }
        status: { type: string, nullable: true }
        requirement_id: { type: string, nullable: true }
        created_at: { type: string }
    CreateTestCaseRequest:
      type: object
      required: [suite_id, title]
      properties:
        suite_id: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: integer }
        status: { type: string }
        requirement_id: { type: string }
    UpdateTestCaseRequest:
      type: object
      properties:
        suite_id: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: integer }
        status: { type: string }
        requirement_id: { type: string }
...
    Execution:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string, nullable: true }
        trigger:
          type: string
          nullable: true
          enum: [manual, ci]
        status:
          type: string
          nullable: true
          enum: [queued, running, completed, failed]
        created_by: { type: string, nullable: true }
        started_at: { type: string, nullable: true }
        ended_at: { type: string, nullable: true }
    CreateExecutionRequest:
      type: object
      required: [project_id, created_by]
      properties:
        project_id: { type: string }
        name: { type: string }
        created_by: { type: string }
    CompleteExecutionRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [completed, failed]
...